fix gcc warnings

+++ src/mod_mrf.cpp
@@ -349,7 +349,7 @@ static int vfile_pread(request_rec *r, storage_manager &mgr,
 
             if ((status != APR_SUCCESS
                     || sr->status != HTTP_PARTIAL_CONTENT
-                    || rctx.size != mgr.size)
+                    || static_cast<size_t>(rctx.size) != mgr.size)
                         && (0 == tries--))
             {
                 ap_log_rerror(APLOG_MARK, APLOG_ERR, 0, r,
@@ -357,7 +357,7 @@ static int vfile_pread(request_rec *r, storage_manager &mgr,
                     name, apr_time_now() - now);
                 failed = true;
             }
-        } while (!failed && rctx.size != mgr.size);
+        } while (!failed && static_cast<size_t>(rctx.size) != mgr.size);
 
         return rctx.size;
     } // Redirect read
@@ -520,9 +520,9 @@ static int handler(request_rec *r) {
 
     tile.l += raster.skip;
     // Check for bad requests, outside of the defined bounds
-    REQ_ERR_IF(tile.l >= raster.n_levels);
+    REQ_ERR_IF(tile.l >= static_cast<size_t>(raster.n_levels));
     rset *level = raster.rsets + tile.l;
-    REQ_ERR_IF(tile.x >= level->w || tile.y >= level->h);
+    REQ_ERR_IF(tile.x >= static_cast<size_t>(level->w) || tile.y >= static_cast<size_t>(level->h));
 
     // Force single z if that's how the MRF is set up, maybe file name mapping applies
     apr_int64_t tz = (raster.size.z != 1) ? tile.z : 0;
@@ -565,7 +565,7 @@ static int handler(request_rec *r) {
     storage_manager img(apr_palloc(r->pool, size), size);
 
     SERR_IF(!img.buffer, "Memory allocation error in mod_mrf");
-    SERR_IF(img.size != vfile_pread(r, img, index.offset, name),
+    SERR_IF(img.size != static_cast<size_t>(vfile_pread(r, img, index.offset, name)),
         "Data read error");
 
     // Looks fine, set the outgoing etag and then the image
